flows/network_speedtest/requirements.txt
prefect>=3.0.0rc17
speedtest-cli